requirements.txt
requests>=2.25.0
beautifulsoup4>=4.9.0
lxml>=4.6.0
python-telegram-bot